Unnecessary cast
git-svn-id: https://svn.apache.org/repos/asf/commons/sandbox/compress/trunk@674244 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/commons/compress/archivers/tar/TarInputStream.java
@@ -307,7 +307,7 @@ public final class TarInputStream extends FilterInputStream
         }
         else
         {
-            return (int)m_oneBuf[ 0 ];
+            return m_oneBuf[ 0 ];
         }
     }
